requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
flake8==7.3.0
flake8-isort==7.0.0
idna==3.11
isort==7.0.0
mccabe==0.7.0
numpy==2.3.5
pandas==2.3.3
pep8-naming==0.15.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pyTelegramBotAPI==4.29.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
telebot==0.0.5
tzdata==2025.3
urllib3==2.5.0
